+++ framework/ui/src/components/organisms/vc-table/vc-table.vue
@@ -1,5 +1,6 @@
 <template>
   <div class="vc-table-wrapper vc-flex vc-flex-column vc-flex-grow_1">
+    <!-- Header slot with filter and searchbar -->
     <slot
       name="header"
       v-if="(items && items.length) || searchValue || searchValue === ''"
@@ -14,6 +15,7 @@
           vc-padding_l
         "
       >
+        <!-- Table filter button -->
         <div class="vc-margin-right_m">
           <vc-table-filter
             :items="filterItems"
@@ -21,64 +23,45 @@
             @reset="$emit('filter:reset')"
           />
         </div>
+
+        <!-- Table search input -->
         <vc-input
           class="vc-flex-grow_1"
           :placeholder="searchPlaceholder"
           :clearable="true"
           :modelValue="searchValue"
-          @update:modelValue="$emit('searchValueChanged', $event)"
+          @update:modelValue="$emit('search:change', $event)"
         ></vc-input>
       </div>
     </slot>
 
     <div class="vc-table-wrapper__inner">
+      <!-- Table loading overlay -->
       <vc-loading :active="loading"></vc-loading>
+
+      <!-- Table scroll container -->
       <vc-container
         v-if="items && items.length"
         ref="scrollContainer"
         :noPadding="true"
         class="vc-flex-grow_1"
       >
-        <template v-if="$isPhone.value && $slots['mobile-item']">
+        <!-- Mobile table view -->
+        <template v-if="$isMobile.value && $slots['mobile-item']">
           <div class="vc-table-mobile">
-            <div
+            <vc-table-mobile-item
               v-for="item in items"
               :key="item.id"
-              class="vc-table-mobile__item"
-              :class="{
-                'vc-table-mobile__item_active':
-                  mobileItems[item.id] && mobileItems[item.id].isActive,
-              }"
-              :style="`transform: translateX(${
-                (mobileItems[item.id] && mobileItems[item.id].offset) || 0
-              }px)`"
+              :item="item"
+              :actionBuilder="itemActionBuilder"
               @click="$emit('itemClick', item)"
-              @touchstart="itemTouchStart($event, item.id)"
-              @touchmove="itemTouchMove($event, item.id)"
-              @touchend="itemTouchEnd($event, item.id)"
-              @touchcancel="itemTouchCancel($event, item.id)"
             >
-              <div class="vc-table-mobile__item-content">
-                <slot name="mobile-item" :item="item"></slot>
-              </div>
-              <div class="vc-table-mobile__item-actions">
-                <div
-                  class="
-                    vc-table-mobile__item-action
-                    vc-table-mobile__item-action_success
-                  "
-                >
-                  <vc-icon icon="fas fa-check" />
-                  <div class="vc-table-mobile__item-action-text">Publish</div>
-                </div>
-                <div class="vc-table-mobile__item-action">
-                  <vc-icon icon="fas fa-ellipsis-h" />
-                  <div class="vc-table-mobile__item-action-text">More</div>
-                </div>
-              </div>
-            </div>
+              <slot name="mobile-item" :item="item"></slot>
+            </vc-table-mobile-item>
           </div>
         </template>
+
+        <!-- Desktop table view -->
         <table
           v-else
           class="vc-table vc-fill_width"
@@ -171,6 +154,8 @@
           </tbody>
         </table>
       </vc-container>
+
+      <!-- Empty table view -->
       <template v-else>
         <slot v-if="searchValue || searchValue === ''" name="notfound">
           <div
@@ -211,6 +196,7 @@
       </template>
     </div>
 
+    <!-- Table footer -->
     <slot name="footer" v-if="items && items.length">
       <div
         class="
@@ -222,12 +208,15 @@
           vc-padding_l
         "
       >
+        <!-- Table pagination -->
         <vc-pagination
           :expanded="expanded"
           :pages="pages"
           :currentPage="currentPage"
           @itemClick="$emit('paginationClick', $event)"
         ></vc-pagination>
+
+        <!-- Table counter -->
         <vc-table-counter
           :label="totalLabel"
           :value="totalCount"
@@ -247,12 +236,7 @@ import VcPagination from "../../molecules/vc-pagination/vc-pagination.vue";
 import VcLoading from "../../atoms/vc-loading/vc-loading.vue";
 import VcTableCounter from "./_internal/vc-table-counter/vc-table-counter.vue";
 import VcTableFilter from "./_internal/vc-table-filter/vc-table-filter.vue";
-
-interface IMobileItem {
-  isActive: boolean;
-  offset: number;
-  start: number;
-}
+import VcTableMobileItem from "./_internal/vc-table-mobile-item/vc-table-mobile-item.vue";
 
 export default defineComponent({
   name: "VcTable",
@@ -266,15 +250,14 @@ export default defineComponent({
     VcTableCounter,
     VcLoading,
     VcTableFilter,
+    VcTableMobileItem,
   },
 
   data() {
     const checkboxes: Record<string, boolean> = {};
-    const mobileItems: Record<string, IMobileItem> = {};
 
     return {
       checkboxes,
-      mobileItems,
     };
   },
 
@@ -292,6 +275,11 @@ export default defineComponent({
     filterItems: {
       type: Array,
       default: () => [],
+    },
+
+    itemActionBuilder: {
+      type: Function,
+      default: undefined,
     },
 
     sort: {
@@ -362,8 +350,10 @@ export default defineComponent({
   emits: [
     "paginationClick",
     "selectionChanged",
-    "searchValueChanged",
+    "search:change",
     "itemClick",
+    "filter:apply",
+    "filter:reset",
   ],
 
   watch: {
@@ -403,30 +393,6 @@ export default defineComponent({
     processCheckbox(id: string, state: boolean) {
       this.checkboxes[id] = state;
       this.$emit("selectionChanged", this.checkboxes);
-    },
-
-    itemTouchStart(e: TouchEvent, id: string): void {
-      this.mobileItems[id] = {
-        ...(this.mobileItems[id] || {}),
-        start: e.touches[0].clientX,
-      };
-    },
-
-    itemTouchMove(e: TouchEvent, id: string): void {
-      const offset = e.touches[0].clientX - this.mobileItems[id].start;
-      if (Math.abs(offset) > 10 && Math.abs(offset) <= 80) {
-        this.mobileItems[id] = {
-          ...(this.mobileItems[id] || {}),
-          offset,
-        };
-      }
-    },
-
-    itemTouchEnd(e: TouchEvent, id: string): void {
-      this.mobileItems[id] = {
-        ...(this.mobileItems[id] || {}),
-        offset: this.mobileItems[id].offset < -50 ? -80 : 0,
-      };
     },
   },
 });
@@ -506,48 +472,6 @@ export default defineComponent({
     &-text {
       font-size: var(--font-size-xl);
       font-weight: var(--font-weight-medium);
-    }
-  }
-
-  &-mobile {
-    &__item {
-      position: relative;
-      display: flex;
-      flex-wrap: nowrap;
-      align-items: stretch;
-
-      &-content {
-        flex-shrink: 0;
-        width: 100%;
-      }
-
-      &-actions {
-        flex-shrink: 0;
-        width: 80px;
-        display: flex;
-        flex-direction: column;
-        justify-content: stretch;
-        background-color: #a9bfd2;
-      }
-
-      &-action {
-        display: flex;
-        flex-grow: 1;
-        flex-basis: 1;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-        color: white;
-
-        &-text {
-          margin-top: 4px;
-          font-size: 14px;
-        }
-
-        &_success {
-          background-color: #87b563;
-        }
-      }
     }
   }
 }