+++ apps/vendor-portal/CHANGELOG.md
@@ -3,6 +3,68 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.51 (2021-10-13)
+
+
+### Features
+
+* pwa sample settings ([64a7f08](https://github.com/VirtoCommerce/platform-manager-sdk/commit/64a7f0818b5bad49f16257d12545fa1f68c32ad9))
+
+
+
+## 0.0.50 (2021-10-12)
+
+
+### Features
+
+* list mobile actions ([c269a88](https://github.com/VirtoCommerce/platform-manager-sdk/commit/c269a88b9ca304520b9157a51ccd4c602fd0059b))
+* mobile swipe transition ([8fdccb1](https://github.com/VirtoCommerce/platform-manager-sdk/commit/8fdccb1345ba3f2d495a6f9344503f1c50b27a7b))
+
+
+
+## 0.0.49 (2021-10-12)
+
+
+### Features
+
+* mobile extended actions, vendor-portal detached ([c92a317](https://github.com/VirtoCommerce/platform-manager-sdk/commit/c92a31743ded9788c939641ccff9922328b6f2f6))
+
+
+
+
+
+## 0.0.50 (2021-10-12)
+
+
+### Features
+
+* list mobile actions ([c269a88](https://github.com/VirtoCommerce/platform-manager-sdk/commit/c269a88b9ca304520b9157a51ccd4c602fd0059b))
+* mobile swipe transition ([8fdccb1](https://github.com/VirtoCommerce/platform-manager-sdk/commit/8fdccb1345ba3f2d495a6f9344503f1c50b27a7b))
+
+
+
+## 0.0.49 (2021-10-12)
+
+
+### Features
+
+* mobile extended actions, vendor-portal detached ([c92a317](https://github.com/VirtoCommerce/platform-manager-sdk/commit/c92a31743ded9788c939641ccff9922328b6f2f6))
+
+
+
+
+
+## 0.0.49 (2021-10-12)
+
+
+### Features
+
+* mobile extended actions, vendor-portal detached ([c92a317](https://github.com/VirtoCommerce/platform-manager-sdk/commit/c92a31743ded9788c939641ccff9922328b6f2f6))
+
+
+
+
+
 ## [0.0.48](https://github.com/VirtoCommerce/platform-manager-sdk/compare/@virtoshell/app-demo-manager@0.0.40...@virtoshell/app-demo-manager@0.0.48) (2021-10-06)